major reorg and cleanup
