+++ src/command_line/mod.rs
@@ -3,7 +3,6 @@ pub mod pubchem;
 pub mod search;
 
 pub mod prelude {
-    pub use clap::{Arg, ArgMatches, Command};
-
     pub use crate::indexing::*;
+    pub use clap::{Arg, ArgMatches, Command};
 }
