+++ src/main.rs
@@ -19,35 +19,20 @@ async fn main() -> eyre::Result<()> {
 
     let app = Command::new("cheminee")
         .subcommand_required(true)
-        .subcommand(command_line::pubchem::fetch_pubchem::command())
-        .subcommand(command_line::indexing::index_sdf::command())
-        .subcommand(command_line::pubchem::stream_pubchem_sdf::command())
-        .subcommand(command_line::search::basic_search::command())
-        .subcommand(rest_api::command())
-        .subcommand(command_line::search::substructure_search::command())
+        .subcommand(command_line::indexing::bulk_index::command())
         .subcommand(command_line::indexing::create_index::command())
         .subcommand(command_line::indexing::delete_index::command())
-        .subcommand(command_line::indexing::bulk_index::command());
+        .subcommand(command_line::indexing::index_sdf::command())
+        .subcommand(command_line::pubchem::fetch_pubchem::command())
+        .subcommand(command_line::pubchem::stream_pubchem_sdf::command())
+        .subcommand(command_line::search::basic_search::command())
+        .subcommand(command_line::search::substructure_search::command())
+        .subcommand(rest_api::command());
 
     let matches = app.get_matches();
     let matches = match matches.subcommand().unwrap() {
-        (command_line::indexing::index_sdf::NAME, matches) => {
-            let writes = command_line::indexing::index_sdf::action(matches)?;
-            log::info!("wrote: {}", writes);
-            Ok(())
-        }
-        (command_line::pubchem::stream_pubchem_sdf::NAME, matches) => {
-            command_line::pubchem::stream_pubchem_sdf::action(matches)
-        }
-        (command_line::search::basic_search::NAME, matches) => {
-            command_line::search::basic_search::action(matches)
-        }
-        (command_line::pubchem::fetch_pubchem::NAME, matches) => {
-            command_line::pubchem::fetch_pubchem::action(matches).await
-        }
-        (rest_api::NAME, matches) => rest_api::action(matches).await,
-        (command_line::search::substructure_search::NAME, matches) => {
-            command_line::search::substructure_search::action(matches)
+        (command_line::indexing::bulk_index::NAME, matches) => {
+            command_line::indexing::bulk_index::action(matches)
         }
         (command_line::indexing::create_index::NAME, matches) => {
             command_line::indexing::create_index::action(matches)
@@ -55,9 +40,24 @@ async fn main() -> eyre::Result<()> {
         (command_line::indexing::delete_index::NAME, matches) => {
             command_line::indexing::delete_index::action(matches)
         }
-        (command_line::indexing::bulk_index::NAME, matches) => {
-            command_line::indexing::bulk_index::action(matches)
+        (command_line::indexing::index_sdf::NAME, matches) => {
+            let writes = command_line::indexing::index_sdf::action(matches)?;
+            log::info!("wrote: {}", writes);
+            Ok(())
         }
+        (command_line::pubchem::fetch_pubchem::NAME, matches) => {
+            command_line::pubchem::fetch_pubchem::action(matches).await
+        }
+        (command_line::pubchem::stream_pubchem_sdf::NAME, matches) => {
+            command_line::pubchem::stream_pubchem_sdf::action(matches)
+        }
+        (command_line::search::basic_search::NAME, matches) => {
+            command_line::search::basic_search::action(matches)
+        }
+        (command_line::search::substructure_search::NAME, matches) => {
+            command_line::search::substructure_search::action(matches)
+        }
+        (rest_api::NAME, matches) => rest_api::action(matches).await,
         (unknown, _) => panic!("🤨: {}", unknown),
     };
