+++ src/rest_api/api/compound_processing/mod.rs
@@ -1,0 +1,5 @@
+mod compound_processing_management;
+pub use compound_processing_management::*;
+
+mod standardize;
+pub use standardize::*;
