+++ src/rest_api/api/compound_processing/standardize.rs
@@ -1,14 +1,36 @@
-use poem_openapi::{payload::Json, ApiResponse, Object};
+use crate::rest_api::api::{StandardizeResponse, StandardizedSmile};
+use crate::{rest_api::models::Smile, search::compound_processing::standardize_smiles};
+use poem_openapi::payload::Json;
 use rayon::prelude::*;
 
-use crate::{rest_api::models::Smile, search::compound_processing::standardize_smiles};
+pub async fn v1_standardize(mol: Json<Vec<Smile>>) -> StandardizeResponse {
+    let standardized_smiles = mol
+        .0
+        .into_par_iter()
+        .map(|s| {
+            let standardize = standardize_smiles(&s.smile);
+
+            match standardize {
+                Ok(romol) => StandardizedSmile {
+                    smile: Some(romol.as_smile()),
+                    error: None,
+                },
+                Err(e) => StandardizedSmile {
+                    smile: Some(s.smile),
+                    error: Some(e.to_string()),
+                },
+            }
+        })
+        .collect::<Vec<_>>();
+
+    StandardizeResponse::Ok(Json(standardized_smiles))
+}
 
 #[cfg(test)]
 mod tests {
-    use poem::{handler, Route};
-
     use super::*;
     use crate::{indexing::index_manager::IndexManager, rest_api::openapi_server::Api};
+    use poem::{handler, Route};
 
     #[handler]
     async fn index() -> StandardizeResponse {
@@ -44,41 +66,4 @@ mod tests {
             .assert_string("CCC=O");
         println!("{:?}", json_value);
     }
-}
-
-#[derive(ApiResponse)]
-pub enum StandardizeResponse {
-    #[oai(status = "200")]
-    Ok(Json<Vec<StandardizedSmile>>),
-}
-
-#[derive(Object, Debug)]
-pub struct StandardizedSmile {
-    #[oai(skip_serializing_if_is_none)]
-    pub smile: Option<String>,
-    #[oai(skip_serializing_if_is_none)]
-    pub error: Option<String>,
-}
-
-pub async fn v1_standardize(mol: Json<Vec<Smile>>) -> StandardizeResponse {
-    let standardized_smiles = mol
-        .0
-        .into_par_iter()
-        .map(|s| {
-            let standardize = standardize_smiles(&s.smile);
-
-            match standardize {
-                Ok(romol) => StandardizedSmile {
-                    smile: Some(romol.as_smile()),
-                    error: None,
-                },
-                Err(e) => StandardizedSmile {
-                    smile: Some(s.smile),
-                    error: Some(e.to_string()),
-                },
-            }
-        })
-        .collect::<Vec<_>>();
-
-    StandardizeResponse::Ok(Json(standardized_smiles))
 }
