+++ src/rest_api/api/indexing/mod.rs
@@ -3,3 +3,15 @@ pub use bulk_index::*;
 
 mod create_index;
 pub use create_index::*;
+
+mod get_index;
+pub use get_index::*;
+
+mod index_management;
+pub use index_management::*;
+
+mod list_indexes;
+pub use list_indexes::*;
+
+mod list_schemas;
+pub use list_schemas::*;
