+++ src/rest_api/api/search/basic_search.rs
@@ -1,7 +1,6 @@
 use crate::indexing::index_manager::IndexManager;
 use crate::rest_api::api::{GetQuerySearchResponse, QueryResponseError};
-use crate::search::aggregate_query_hits;
-use crate::search::basic_search::basic_search;
+use crate::search::{aggregate_query_hits, basic_search::basic_search};
 use poem_openapi::payload::Json;
 
 pub fn v1_index_search_basic(
