+++ src/rest_api/api/search/mod.rs
@@ -1,39 +1,8 @@
 mod basic_search;
-mod substructure_search;
-
 pub use basic_search::*;
-use poem_openapi::payload::Json;
-use poem_openapi_derive::{ApiResponse, Object};
+
+mod search_management;
+pub use search_management::*;
+
+mod substructure_search;
 pub use substructure_search::*;
-
-use crate::search::{QuerySearchHit, StructureSearchHit};
-
-#[derive(ApiResponse)]
-pub enum GetQuerySearchResponse {
-    #[oai(status = "200")]
-    Ok(Json<Vec<QuerySearchHit>>),
-    #[oai(status = "404")]
-    IndexDoesNotExist,
-    #[oai(status = "500")]
-    Err(Json<QueryResponseError>),
-}
-
-#[derive(Object, Debug)]
-pub struct QueryResponseError {
-    pub error: String,
-}
-
-#[derive(ApiResponse)]
-pub enum GetStructureSearchResponse {
-    #[oai(status = "200")]
-    Ok(Json<Vec<StructureSearchHit>>),
-    #[oai(status = "404")]
-    IndexDoesNotExist,
-    #[oai(status = "500")]
-    Err(Json<StructureResponseError>),
-}
-
-#[derive(Object, Debug)]
-pub struct StructureResponseError {
-    pub error: String,
-}
